Прописан паттерн MVP для AddAssortment. У цены double заменен на decimal. Убранны лишние using.

+++ CourseWorkWF/Models/Product.cs
@@ -10,13 +10,13 @@ namespace CourseWorkWF
     {
         private int _productID;
         private string _productName;
-        private double _price;
+        private decimal _price;
         private int _amount;
         public int ProductID { get { return _productID; } set { if (value < 0) throw new AccessViolationException("ID продукта отрицательно"); _productID = value; } }
         public string Name { get { return _productName; } set { if (value == null) throw new ArgumentNullException("Имя продукта пустое"); _productName = value; } }
-        public double Price { get { return _price; } set { if (value < 0) throw new AccessViolationException("Цена продукта отрицательна"); _price = value; } }
+        public decimal Price { get { return _price; } set { if (value < 0) throw new AccessViolationException("Цена продукта отрицательна"); _price = value; } }
         public int Amount { get { return _amount; } set { if (value < 0) throw new AccessViolationException("Число продуктов отрицательно"); _amount = value; } }
-        public Product(int productID, string name, double price, int _amount)
+        public Product(int productID, string name, decimal price, int _amount)
         {
             ProductID = productID;
             Name = name;
